% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_stats.R
\name{pairwise_stats}
\alias{pairwise_stats}
\title{Calculate pairwise statistics between groups}
\usage{
pairwise_stats(x, factor_col, var_col, fxn, two_way = FALSE, ...)
}
\arguments{
\item{x}{A data.frame or tibble.}

\item{factor_col}{Name of the grouping column}

\item{var_col}{Name of the variable of interest}

\item{fxn}{The function to be applied}

\item{two_way}{Whether the order of data inputs to fxn matter}

\item{...}{Extra arguments passed to fxn, i.e. alternative = "greater", etc.}
}
\value{
A data.frame
}
\description{
pairwise_stats performs a provided function pairwise between all combinations of groups. The first two arguments of the function passed to pairwise_stats must accept vectors of values as inputs. These vectors should correspond to the values for group A and group B, respectively.

The function takes as input a data.frame or tibble, the column name of a categorical variable, the column name for a variable of interest, and a function.
}
