% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summarize.R
\name{print.group_summary}
\alias{print.group_summary}
\title{Print a group_summary object}
\usage{
\method{print}{group_summary}(x, num_to_print = 3, ...)
}
\arguments{
\item{x}{An object of class group_summary.}

\item{num_to_print}{An integer specifying the number of tibbles to print.}

\item{...}{Additional parameters passed to print.}
}
\description{
Print a group_summary object
}
\examples{
group_summarize(iris, "Species", c("Sepal.Length", "Sepal.Width"))
}
