% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_stats.R
\name{print.pairwise_stats}
\alias{print.pairwise_stats}
\title{Print a pairwise_stats object}
\usage{
\method{print}{pairwise_stats}(x, ...)
}
\arguments{
\item{x}{An object of class pairwise_stats.}

\item{...}{Additional parameters passed to print.}
}
\description{
Print a pairwise_stats object
}
\examples{
\dontrun{
print(pairwise_stats(iris, "Species", "Sepal.Length", t.test))
}
}
